refactor dir name pydantic -> models
