+++ geoips/models/sectors.py
@@ -12,7 +12,7 @@ from pydantic import (
 )
 from pydantic.functional_validators import AfterValidator
 
-from geoips.pydantic.bases import FrozenModel, PermissiveFrozenModel, PluginModel
+from geoips.models.bases import FrozenModel, PermissiveFrozenModel, PluginModel
 
 
 class EarthConstants(float, Enum):
