+++ tests/unit_tests/models/sectors/test_cases.yaml
