+++ tests/unit_tests/plugins/yaml/test_all_yaml_plugins.py
@@ -8,7 +8,7 @@ from importlib import resources, metadata
 
 from geoips.interfaces.base import YamlPluginValidator
 from geoips.interfaces.yaml_based.products import ProductsPluginValidator
-from geoips.pydantic.workflows import WorkflowPluginModel
+from geoips.models.workflows import WorkflowPluginModel
 
 validator = YamlPluginValidator()
 product_validator = ProductsPluginValidator()
